+++ Changes.md
@@ -1,3 +1,4 @@
+* #1000 - Add report of regular match play runs for the refs
 * #1012 - Remove need for jQuery validation library
 * #958 - Replace jQuery datetime picker and time picker with standard input types
 * #998 - Scale the main scoreboard based on screen size